+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `make_website_yaml` now takes a dataframe (the output of `get_deployment_table`) as an input, rather than a database connection
     - The `make...yaml` functions now return the dictionary rather than the file path
     - `_read_esdglider_yaml` is now a full internal function
+    - Added ability to use new Location table columns (Region_Name and Sea_Name)
 - Changed plots to use a set legend position when possible, and to remove the 'inflections' tvt plot as it is functionally a duplicate of the 'diveEnergy' tvt plot
 - Changed `glider.binary_to_nc` to make the gridded data files using `glider.make_gridfiles_depth_measured` if the argument 'sci_timeseries_pyglider' is True
 - Changed `glider` so that science post-processing happens consistently in `glider.binary_to_nc`, rather than `glider.timeseries_raw_to_sci`